+++ modelo/conexion.php
@@ -1,7 +1,7 @@
 <?php
     $conexion = mysqli_connect("localhost", 
-        "Mauricio", 
-        "maury18", 
+        "root", 
+        "", 
         "compostela") or 
     die("Problemas con la conexión");
 ?>